Temp Commit
----------------------------------------
Toast are included to avoid check the application
activities in all the mobiles.

This update was initiated since it didn't work
on one of out Tester's mobile.

Remark: Still the app is not working in the
tester's mobile. Either it is the problem of our
app, or the problem in the mobile.

Let's look further and pray that the fault is
on the tester's mobile.

Let's work further and handover the job to the
client as soon as possible.
----------------------------------------
* UPDATE RESULT : FAILED (1%)
* SUCCESS RATIO : PASSED (99%)

* TOTAL OUTCOME : 99% (till date)

----------------------------------------